add getters into classes in order to fetch data and making endpoints working

+++ microservice-listing/src/main/java/org/efrei/DAO/ListingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ListingRepository extends JpaRepository<Listing, Long> {
+public interface ListingRepository extends JpaRepository<Listing, Integer> {
     List<Listing> findByAvailability(Boolean availability);
 
     Listing findByTitle(String title);
